+++ packages/button/src/index.ts
@@ -1,1 +1,3 @@
 export * from './lib/Button';
+export * from './lib/IconButton';
+export * from './lib/LoadingButton';