+++ packages/paper/src/lib/Paper.tsx
@@ -1,7 +1,13 @@
 import { forwardRef } from 'react';
 import { Paper as MuiPaper, PaperProps as MuiPaperProps } from '@mui/material';
 
-export type PaperProps = Omit<MuiPaperProps, 'classes' | 'style' | 'className' | 'sx'>;
+export type PaperProps = Omit<MuiPaperProps, 'classes' | 'style' | 'className' | 'sx' | 'variant'> & {
+  /**
+   * The variant to use.
+   * @default 'outlined'
+   */
+  variant?: 'elevation' | 'outlined';
+};
 
 export const Paper = forwardRef<HTMLDivElement, PaperProps>((props, ref) => {
   return <MuiPaper {...props} ref={ref} />;