+++ packages/theme/src/lib/light-theme.ts
@@ -281,85 +281,59 @@ export const lightTheme = {
         disableElevation: true,
       },
       styleOverrides: {
-        root: {
-          minWidth: '128px',
-          textTransform: 'none',
-        },
+        root: ({ ownerState } : any) => ({
+          ...(ownerState.variant === 'contained' ?
+            {
+              backgroundColor: tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Main` as keyof typeof tokens],
+              color: tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Contrast` as keyof typeof tokens],
+              "&:hover": {
+                backgroundColor: tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Dark` as keyof typeof tokens],
+                boxShadow: 'none',
+              },
+              "&:focus": {
+                backgroundColor: tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Dark` as keyof typeof tokens],
+                outline: '2px solid white',
+                boxShadow: `0 0 0px 4px ${tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Dark` as keyof typeof tokens]}`,
+              },
+              "&:active": {
+                backgroundColor: tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Main` as keyof typeof tokens],
+              }
+            }
+            : {
+              "&:hover": {
+                boxShadow: 'none',
+              },
+              "&:focus": {
+                outline: '2px solid white',
+                boxShadow: `0 0 0px 4px ${tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Main` as keyof typeof tokens]}`,
+              },
+            }
+          ),
+          boxShadow: 'none',
+          fontWeight: tokens.fontWeightsBold,
+          textTransform: 'none'
+        }),
       },
-      variants: [
-        {
-          props: { color: 'tertiary', variant: 'contained' },
-          style: {
-            backgroundColor: tokens.colorTertiaryMain,
-            color: tokens.colorTertiaryContrast,
-            '&:hover': {
-              backgroundColor: tokens.colorTertiaryDark,
-            },
-            '&:focus': {
-              backgroundColor: tokens.colorTertiaryDark,
-              outline: '2px solid white',
-              boxShadow: '0 0 0px 4px rgba(0,0,0,1)',
-            },
-            '&:active': {
-              backgroundColor: tokens.colorTertiaryMain,
-            },
+    },
+    MuiIconButton: {
+      defaultProps: {
+        // The props to change the default for.
+        disableRipple: true, // No more ripple, on the whole application 💣!
+        disableElevation: true
+      },
+      styleOverrides: {
+        root: {
+          color: tokens.colorTextPrimary,
+          "&:hover": {
+            backgroundColor: tokens.colorTertiaryDark,
           },
+          "&:focus": {
+            backgroundColor: tokens.colorTertiaryDark,
+            outline: `2px solid ${tokens.colorCommonWhite}`,
+            boxShadow: `0 0 0px 4px ${tokens.colorSecondaryLight}`,
+          }
         },
-        {
-          props: { color: 'secondary' },
-          style: {
-            backgroundColor: tokens.colorSecondaryMain,
-            color: tokens.colorSecondaryContrast,
-            '&:hover': {
-              backgroundColor: tokens.colorSecondaryDark,
-            },
-            '&:focus': {
-              backgroundColor: tokens.colorSecondaryDarker,
-              outline: '2px solid white',
-              boxShadow: '0 0 0px 4px rgba(0,0,0,1)',
-            },
-            '&:active': {
-              backgroundColor: tokens.colorSecondaryMain,
-            },
-          },
-        },
-        {
-          props: { color: 'primary' },
-          style: {
-            backgroundColor: tokens.colorPrimaryMain,
-            color: tokens.colorPrimaryContrast,
-            '&:hover': {
-              backgroundColor: tokens.colorPrimaryDark,
-            },
-            '&:focus': {
-              backgroundColor: tokens.colorPrimaryMain,
-              outline: '2px solid white',
-              boxShadow: '0 0 0px 4px rgba(21,82,188,1)',
-            },
-            '&:active': {
-              backgroundColor: tokens.colorPrimaryLight,
-            },
-          },
-        },
-        {
-          props: { color: 'error', variant: 'contained' },
-          style: {
-            backgroundColor: tokens.colorErrorMain,
-            color: tokens.colorErrorContrast,
-            '&:hover': {
-              backgroundColor: tokens.colorErrorDark,
-            },
-            '&:focus': {
-              backgroundColor: tokens.colorErrorMain,
-              outline: '2px solid white',
-              boxShadow: '0 0 0px 4px rgba(21,82,188,1)',
-            },
-            '&:active': {
-              backgroundColor: tokens.colorErrorLight,
-            },
-          },
-        },
-      ],
+      }
     },
     MuiLink: {
       defaultProps: {
@@ -374,13 +348,6 @@ export const lightTheme = {
           '&:visited': {
             color: tokens.colorIndigo600,
           },
-        },
-      },
-    },
-    MuiLoadingButton: {
-      styleOverrides: {
-        root: {
-          position: 'relative',
         },
       },
     },
@@ -401,5 +368,15 @@ export const lightTheme = {
         },
       },
     },
+    MuiTooltip: {
+      styleOverrides: {
+        tooltip: {
+          backgroundColor: tokens.colorGrey400,
+          fontSize: '.785rem',
+          fontStyle: 'normal',
+          padding: '2px 8px'
+        },
+      },
+    }
   },
 };