+++ packages/theme/src/lib/theme.stories.tsx
@@ -1,4 +1,4 @@
-import { lightTheme } from '@availity/theme';
+import { lightTheme } from './light-theme';
 
 export const Default = {
   render: () => <pre>{JSON.stringify(lightTheme, null, 2)}</pre>,