+++ packages/paper/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-paper@0.1.0...@availity/mui-paper@0.1.1) (2023-05-23)
+
 ## 0.1.0 (2023-05-22)